fix: rename search article component
